Use same code for windows and unix docker diskio
